requirements.txt
Pillow==10.0.0
tkmacosx==1.0.5